Fix microservices startup issues and enable Swagger UI for local development
- Fix bean definition conflicts by adding allow-bean-definition-overriding
- Resolve circular dependencies with allow-circular-references
- Fix Flyway migrations for H2 compatibility:
  - Split multi-column ALTER TABLE statements
  - Move inline INDEX definitions to separate CREATE INDEX statements
  - Fix duplicate migration versions (V9, V3) by renaming
  - Change DECIMAL to DOUBLE PRECISION for H2 compatibility
  - Add missing tables (refund_policies, refund_status)
- Simplify security configuration for local dev:
  - Disable OAuth2 JWT validation at API Gateway
  - Enable simple security for user service
  - Remove JwtDecoder dependencies from filters
- Fix Swagger UI configuration:
  - Add springdoc.swagger-ui.url to all services
  - Configure correct API docs paths
- Remove duplicate @FilterDef annotations from entities
- Add PasswordEncoderConfig for user service
- Fix entity field types (remove precision/scale from Double fields)
- Disable SimpleNotificationController by default to avoid conflicts

+++ irctc-train-service/src/main/resources/db/migration/V1__Initial_train_schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS trains (
   arrival_time TIMESTAMP NOT NULL,
   train_type VARCHAR(50) NOT NULL,
   train_class VARCHAR(50) NOT NULL,
-  base_fare DECIMAL(10, 2) NOT NULL,
+  base_fare DOUBLE PRECISION NOT NULL,
   total_seats INT NOT NULL,
   available_seats INT NOT NULL,
   status VARCHAR(50) NOT NULL,
